# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BeePlan API
  version: "0.1.0"
servers:
  - url: http://localhost:8000
paths:
  /v1/health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
  /v1/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "200": { description: Created user }
  /v1/auth/token:
    post:
      summary: Login (password grant style)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: JWT access token
  /v1/me:
    get:
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        "200": { description: User }
  /v1/apiaries:
    get:
      summary: List apiaries
      security:
        - bearerAuth: []
      responses:
        "200": { description: List }
    post:
      summary: Create apiary
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "201": { description: Created apiary }
  /v1/apiaries/{apiary_id}:
    patch:
      summary: Update apiary
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: apiary_id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200": { description: Updated apiary }
    delete:
      summary: Delete apiary
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: apiary_id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
  /v1/bee-breeds:
    get:
      summary: List bee breeds (read-only catalog)
      security:
        - bearerAuth: []
      responses:
        "200": { description: List of breeds }
  /v1/colonies/suggested-name:
    get:
      summary: Suggest auto-generated colony name
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Suggested name
          content:
            application/json:
              schema:
                type: object
                required: [name]
                properties:
                  name: { type: string, example: "Альбина 034" }
  /v1/colonies:
    get:
      summary: List colonies for apiary
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: apiary_id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: List }
    post:
      summary: Create colony
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiary_id, name]
              properties:
                apiary_id: { type: integer }
                name: { type: string }
                bee_breed: { type: string, nullable: true }
                description: { type: string, nullable: true }
                colony_type: { type: string, nullable: true, enum: [split, nucleus, colony, swarm] }
                hive_type: { type: string, nullable: true }
                body_count: { type: integer, nullable: true }
                frames_per_body: { type: integer, nullable: true }
                hive_volume_m3: { type: number, nullable: true }
      responses:
        "201": { description: Created colony }
  /v1/colonies/{colony_id}:
    get:
      summary: Get colony
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: colony_id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Colony }
    patch:
      summary: Update colony
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: colony_id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                bee_breed: { type: string, nullable: true }
                description: { type: string, nullable: true }
                colony_type: { type: string, nullable: true, enum: [split, nucleus, colony, swarm] }
                hive_type: { type: string, nullable: true }
                body_count: { type: integer, nullable: true }
                frames_per_body: { type: integer, nullable: true }
                hive_volume_m3: { type: number, nullable: true }
      responses:
        "200": { description: Updated colony }
    delete:
      summary: Delete colony
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: colony_id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
  /v1/concentrators:
    get:
      summary: List concentrators for apiary
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: apiary_id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: List }
    post:
      summary: Create concentrator
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiary_id, name]
              properties:
                apiary_id: { type: integer }
                name: { type: string }
      responses:
        "201": { description: Created concentrator with ingest_token }
  /v1/concentrators/{concentrator_id}:
    patch:
      summary: Update concentrator
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: concentrator_id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200": { description: Updated concentrator }
    delete:
      summary: Delete concentrator
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: concentrator_id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
  /v1/edge-devices:
    get:
      summary: List edge devices for apiary
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: apiary_id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: List }
    post:
      summary: Create edge device
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [concentrator_id, public_id]
              properties:
                concentrator_id: { type: integer }
                public_id: { type: string }
                label: { type: string, nullable: true }
                colony_id: { type: integer, nullable: true }
      responses:
        "201": { description: Created device }
  /v1/edge-devices/{device_id}:
    patch:
      summary: Update edge device
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: device_id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                public_id: { type: string }
                label: { type: string, nullable: true }
      responses:
        "200": { description: Updated device }
    delete:
      summary: Delete edge device
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: device_id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
  /v1/edge-devices/{device_id}/colony:
    put:
      summary: Set active colony for device (writes assignment history)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: device_id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                colony_id: { type: integer, nullable: true }
      responses:
        "200": { description: Updated device }
  /v1/telemetry/batch:
    post:
      summary: Ingest telemetry samples from concentrator
      security:
        - concentratorBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [samples]
              properties:
                samples:
                  type: array
                  items:
                    type: object
                    required: [device_public_id, metric, ts, value]
                    properties:
                      device_public_id: { type: string }
                      metric: { type: string }
                      ts: { type: string, format: date-time }
                      value: {}
      responses:
        "200": { description: Batch result }
  /v1/colonies/{colony_id}/telemetry:
    get:
      summary: Query telemetry for colony
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: colony_id
          required: true
          schema: { type: integer }
        - in: query
          name: metric
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 500 }
        - in: query
          name: from
          schema: { type: string, format: date-time }
          description: Inclusive start of time range
        - in: query
          name: to
          schema: { type: string, format: date-time }
          description: Inclusive end of time range
      responses:
        "200": { description: Points }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    concentratorBearer:
      type: http
      scheme: bearer
      description: Use concentrator ingest_token as Bearer value